+++ core/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 1.53.0
 ### Breaking Changes
 * Revert `StartService` and `LoadModule` changes from 1.52.0